==> 01_QC/03_filteredCells.Rmd <==
---
title: "Subsetting Cells"
author: "kou"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, include = T)
```

## "Finalizing" filtered cells for downstream cells
Want to subset cells that passed Seurat (RNA) and ArchR (ATAC) filtering.
```{r load}
library(Seurat)
library(ArchR)
library(tidyverse)
```

```{r dataImport}
seur_flt = qread("./01_QC/rds/01_seur_filt.qs")
archR_flt = loadArchRProject("./01_QC/ArchRproj_flt")

# merge seurat, get metadata
seur_flt = merge(seur_flt[[1]], 
                 seur_flt[-1],
                 add.cell.ids = names(seur_flt))

# get cell names
atac_cells = archR_flt$cellNames
rna_cells = Cells(seur_flt)

# get intersection
keep_ids = intersect(atac_cells, gsub("_","#",rna_cells))

# subset
seur_keep = seur_flt[,rna_cells %in% gsub("#","_",keep_ids)]
qsave(seur_keep, "./01_QC/rds/03_seur_keep.qs", nthreads=16)

# TO DO: use subsetArchRProject() to drop cells in arrow files!
archR_keep = archR_flt[archR_flt$cellNames %in% keep_ids,]
saveArchRProject(ArchRProj = archR_keep, 
                             outputDirectory = "01_QC/archRprojects/ArchRproj_keep", 
                             load = F)
```
